refactor: extract helper function to reduce code duplication
The only diff between `getContainerBuildActionOutputs` and `k8sGetContainerBuildActionOutputs`
was in the deploymentImageId calculation that depended
on the presence of deployment registry config.
